+++ src/Archive/Formats/WadJArchive.h
@@ -16,7 +16,7 @@ public:
 	bool write(MemChunk& mc, bool update = true) override; // Write to MemChunk
 
 	string detectNamespace(ArchiveEntry* entry) override;
-	string detectNamespace(size_t index, ArchiveDir* dir = nullptr) override;
+	string detectNamespace(unsigned index, ArchiveDir* dir = nullptr) override;
 
 	static bool isWadJArchive(MemChunk& mc);
 	static bool isWadJArchive(const string& filename);
